refactor(modeler): change language to typescript
