+++ apps/miranum-modeler-webview/.eslintrc.json
@@ -8,7 +8,9 @@
     },
     {
       "files": ["*.ts", "*.tsx"],
-      "rules": {}
+      "rules": {
+        "@typescript-eslint/no-explicit-any": "warn"
+      }
     },
     {
       "files": ["*.js", "*.jsx"],
